// StripePoc.React/src/models/foundations/payments/SubscriptionPayload.ts
// ---------------------------------------------------------------
// Copyright (c) Emerald Kilonova. All rights reserved.
// ---------------------------------------------------------------

export class SubscriptionPayload {
    accountId: string;
    serviceName: string;
    amount: number;
    currency: string;
    stripePriceId: string;
}