// StripePoc.React/src/models/foundations/payments/OneTimePaymentPayload.ts
// ---------------------------------------------------------------
// Copyright (c) Emerald Kilonova. All rights reserved.
// ---------------------------------------------------------------

export class OneTimePaymentPayload {
    accountId: string;
    amount: number;
    currency: string;
    status: number;
}