// StripePoc.React/src/models/foundations/payments/PaymentView.ts
// ---------------------------------------------------------------
// Copyright (c) Emerald Kilonova. All rights reserved.
// ---------------------------------------------------------------

export class PaymentView {
    id: string;
    amount: number;
    currency: string;
    status: string;
    occurredAt: string;
}